+++ mayan/apps/cabinets/models.py
@@ -86,20 +86,20 @@ class Cabinet(ExtraDataModelMixin, MPTTModel):
             permission=permission_document_view, user=user
         ).count()
 
-    def get_documents_queryset(self, permission=None, user=None):
+    def get_documents_queryset(self, user, permission=None):
         """
         Provide a queryset of the documents in a cabinet. The queryset is
         filtered by access.
         """
-        queryset = self.documents.all()
+        queryset = Document.valid.filter(pk__in=self.documents.all())
 
         if permission and user:
             queryset = AccessControlList.objects.restrict_queryset(
-                permission=permission_document_view, queryset=queryset,
+                permission=permission, queryset=queryset,
                 user=user
             )
 
-        return Document.valid.filter(pk__in=queryset.values('pk'))
+        return queryset
 
     def get_full_path(self):
         """