+++ HISTORY.rst
@@ -218,6 +218,8 @@
 - Don't introspect document file MIME type at download. Instead pass the
   stored values.
 - Filter trashed documents from the tag document count column.
+- Filter trashed documents from the cabinet document retrieval method. This
+  brings code parity with tags which work in a very similar way.
 
 4.2.3 (2022-04-01)
 ==================